+++ packages/atproto_core/test/src/sessions/sessions_test.dart
@@ -92,4 +92,42 @@ void main() {
       );
     });
   });
+
+  group('.deleteSession', () {
+    test('normal case', () async {
+      final response = await deleteSession(
+        refreshJwt: 'xxxxxx',
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/sessions/data/delete_session.json',
+        ),
+      );
+
+      expect(response, isA<xrpc.XRPCResponse>());
+      expect(response.data, isA<xrpc.EmptyData>());
+    });
+
+    test('when unauthorized', () async {
+      atp_test.expectUnauthorizedException(
+        () async => await deleteSession(
+          refreshJwt: 'xxxxxx',
+          mockedPostClient: atp_test.createMockedPostClient(
+            'test/src/sessions/data/error.json',
+            statusCode: 401,
+          ),
+        ),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      atp_test.expectRateLimitExceededException(
+        () async => await deleteSession(
+          refreshJwt: 'xxxxxx',
+          mockedPostClient: atp_test.createMockedPostClient(
+            'test/src/sessions/data/error.json',
+            statusCode: 429,
+          ),
+        ),
+      );
+    });
+  });
 }